refactor: separate server layers
