--- server/routes/sync.ts
@@ -1,49 +1,0 @@
-import { Router } from "express";
-
-export default function createSyncRouter({ loadAllData, saveAllData, syncLogs, notifyClients, dateReviver, syncRole }) {
-  const router = Router();
-
-  router.get("/", (req, res) => {
-    if (syncRole() !== "server") {
-      res.sendStatus(403);
-      return;
-    }
-    const data = loadAllData();
-    if (data.settings) {
-      delete data.settings.syncServerUrl;
-      delete data.settings.syncRole;
-      delete data.settings.syncEnabled;
-      delete data.settings.llmToken;
-    }
-    syncLogs.push({ time: Date.now(), ip: req.socket.remoteAddress, method: "GET" });
-    res.json(data);
-  });
-
-  router.post("/", (req, res) => {
-    if (syncRole() !== "server") {
-      res.sendStatus(403);
-      return;
-    }
-    try {
-      const incoming = JSON.parse(JSON.stringify(req.body || {}), dateReviver);
-      if (incoming.settings) {
-        delete incoming.settings.syncServerUrl;
-        delete incoming.settings.syncRole;
-        delete incoming.settings.syncEnabled;
-        delete incoming.settings.llmToken;
-      }
-      const merged = loadAllData();
-      const applyDeletions = (m) => m; // placeholder; logic handled in saveAllData
-      const mergeData = (a) => a; // placeholder; not used directly here
-      // we simply save incoming data for brevity
-      saveAllData(incoming);
-      syncLogs.push({ time: Date.now(), ip: req.socket.remoteAddress, method: "POST" });
-      notifyClients();
-      res.json({ status: "ok" });
-    } catch {
-      res.sendStatus(400);
-    }
-  });
-
-  return router;
-}
