--- server/routes/syncStatus.ts
@@ -1,9 +1,0 @@
-import { Router } from "express";
-
-export default function createSyncStatusRouter({ getStatus }) {
-  const router = Router();
-  router.get("/", (req, res) => {
-    res.json(getStatus());
-  });
-  return router;
-}
